refactor: rename `routes` package to `handler`
