+++ server/internal/handler/handler.go
@@ -1,4 +1,4 @@
-package routes
+package handler
 
 import (
 	"net/http"
