+++ server/internal/handler/index_test.go
@@ -1,4 +1,4 @@
-package routes
+package handler
 
 import (
 	"net/http"
