+++ server/internal/handler/otp.go
@@ -1,4 +1,4 @@
-package routes
+package handler
 
 import (
 	"bytes"
